Create a theme provider component
